requirements.txt
fastapi
uvicorn
openpyxl
jinja2
python-multipart
pandas
psycopg2-binary
